Refactor FXIOS-6597-6596 [v116] Reorganize about and support settings (#14873)
* Reorganize about settings

* Reorganize privacy settings
