+++ Client/Frontend/Settings/Main/About/YourRightsSetting.swift
@@ -1,0 +1,21 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+/// Opens about:rights page in the content view controller
+class YourRightsSetting: Setting {
+    override var title: NSAttributedString? {
+        return NSAttributedString(string: .AppSettingsYourRights,
+                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
+    }
+
+    override var url: URL? {
+        return URL(string: "https://www.mozilla.org/about/legal/terms/firefox/")
+    }
+
+    override func onClick(_ navigationController: UINavigationController?) {
+        setUpAndPushSettingsContentViewController(navigationController, self.url)
+    }
+}
