--- Client/Frontend/Settings/Main/AppSettingsOptions.swift
@@ -1,225 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/
-
-import Common
-import Foundation
-import Shared
-import Account
-import LocalAuthentication
-
-// Show the current version of Firefox
-class VersionSetting: Setting {
-    unowned let settings: SettingsTableViewController
-
-    override var accessibilityIdentifier: String? { return "FxVersion" }
-    weak var appSettingsDelegate: AppSettingsDelegate?
-
-    init(settings: SettingsTableViewController,
-         appSettingsDelegate: AppSettingsDelegate) {
-        self.settings = settings
-        self.appSettingsDelegate = appSettingsDelegate
-        super.init(title: nil)
-    }
-
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: "\(AppName.shortName) \(AppInfo.appVersion) (\(AppInfo.buildNumber))", attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override func onConfigureCell(_ cell: UITableViewCell, theme: Theme) {
-        super.onConfigureCell(cell, theme: theme)
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        appSettingsDelegate?.clickedVersion()
-    }
-
-    override func onLongPress(_ navigationController: UINavigationController?) {
-        copyAppVersionAndPresentAlert(by: navigationController)
-    }
-
-    func copyAppVersionAndPresentAlert(by navigationController: UINavigationController?) {
-        let alertTitle: String = .SettingsCopyAppVersionAlertTitle
-        let alert = AlertController(title: alertTitle, message: nil, preferredStyle: .alert)
-        getSelectedCell(by: navigationController)?.setSelected(false, animated: true)
-        UIPasteboard.general.string = self.title?.string
-        navigationController?.topViewController?.present(alert, animated: true) {
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
-                alert.dismiss(animated: true)
-            }
-        }
-    }
-
-    func getSelectedCell(by navigationController: UINavigationController?) -> UITableViewCell? {
-        let controller = navigationController?.topViewController
-        let tableView = (controller as? AppSettingsTableViewController)?.tableView
-        guard let indexPath = tableView?.indexPathForSelectedRow else { return nil }
-        return tableView?.cellForRow(at: indexPath)
-    }
-}
-
-// Opens the license page in a new tab
-class LicenseAndAcknowledgementsSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsLicenses, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "\(InternalURL.baseUrl)/\(AboutLicenseHandler.path)")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-// Opens the App Store review page of this app
-class AppStoreReviewSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .Settings.About.RateOnAppStore, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        RatingPromptManager.goToAppStoreReview()
-    }
-}
-
-// Opens about:rights page in the content view controller
-class YourRightsSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsYourRights,
-                                  attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "https://www.mozilla.org/about/legal/terms/firefox/")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-// Opens the on-boarding screen again
-class ShowIntroductionSetting: Setting {
-    let profile: Profile
-
-    override var accessibilityIdentifier: String? { return "ShowTour" }
-
-    init(settings: SettingsTableViewController) {
-        self.profile = settings.profile
-        let attributes = [NSAttributedString.Key.foregroundColor: settings.themeManager.currentTheme.colors.textPrimary]
-        super.init(title: NSAttributedString(string: .AppSettingsShowTour,
-                                             attributes: attributes))
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        navigationController?.dismiss(animated: true, completion: {
-            NotificationCenter.default.post(name: .PresentIntroView, object: self)
-
-            TelemetryWrapper.recordEvent(
-                category: .action,
-                method: .tap,
-                object: .settingsMenuShowTour
-            )
-        })
-    }
-}
-
-class SendFeedbackSetting: Setting {
-    override var title: NSAttributedString? {
-        return NSAttributedString(string: .AppSettingsSendFeedback, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary])
-    }
-
-    override var url: URL? {
-        return URL(string: "https://connect.mozilla.org/")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-class SendAnonymousUsageDataSetting: BoolSetting {
-    init(prefs: Prefs, delegate: SettingsDelegate?, theme: Theme) {
-        let statusText = NSMutableAttributedString()
-        statusText.append(NSAttributedString(string: .SendUsageSettingMessage, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textSecondary]))
-        statusText.append(NSAttributedString(string: " "))
-        statusText.append(NSAttributedString(string: .SendUsageSettingLink, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.actionPrimary]))
-
-        super.init(
-            prefs: prefs,
-            prefKey: AppConstants.prefSendUsageData,
-            defaultValue: true,
-            attributedTitleText: NSAttributedString(string: .SendUsageSettingTitle),
-            attributedStatusText: statusText,
-            settingDidChange: {
-                AdjustHelper.setEnabled($0)
-                DefaultGleanWrapper.shared.setUpload(isEnabled: $0)
-                Experiments.setTelemetrySetting($0)
-            }
-        )
-        // We make sure to set this on initialization, in case the setting is turned off
-        // in which case, we would to make sure that users are opted out of experiments
-        Experiments.setTelemetrySetting(prefs.boolForKey(AppConstants.prefSendUsageData) ?? true)
-    }
-
-    override var accessibilityIdentifier: String? { return "SendAnonymousUsageData" }
-
-    override var url: URL? {
-        return SupportUtils.URLForTopic("adjust")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-class StudiesToggleSetting: BoolSetting {
-    init(prefs: Prefs, delegate: SettingsDelegate?, theme: Theme) {
-        let statusText = NSMutableAttributedString()
-        statusText.append(NSAttributedString(string: .SettingsStudiesToggleMessage, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textSecondary]))
-        statusText.append(NSAttributedString(string: " "))
-        statusText.append(NSAttributedString(string: .SettingsStudiesToggleLink, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.actionPrimary]))
-
-        super.init(
-            prefs: prefs,
-            prefKey: AppConstants.prefStudiesToggle,
-            defaultValue: true,
-            attributedTitleText: NSAttributedString(string: .SettingsStudiesToggleTitle),
-            attributedStatusText: statusText,
-            settingDidChange: {
-                Experiments.setStudiesSetting($0)
-            }
-        )
-        // We make sure to set this on initialization, in case the setting is turned off
-        // in which case, we would to make sure that users are opted out of experiments
-        Experiments.setStudiesSetting(prefs.boolForKey(AppConstants.prefStudiesToggle) ?? true)
-    }
-
-    override var accessibilityIdentifier: String? { return "StudiesToggle" }
-
-    override var url: URL? {
-        return SupportUtils.URLForTopic("ios-studies")
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        setUpAndPushSettingsContentViewController(navigationController, self.url)
-    }
-}
-
-// Opens the SUMO page in a new tab
-class OpenSupportPageSetting: Setting {
-    init(delegate: SettingsDelegate?, theme: Theme) {
-        super.init(title: NSAttributedString(string: .AppSettingsHelp, attributes: [NSAttributedString.Key.foregroundColor: theme.colors.textPrimary]),
-                   delegate: delegate)
-    }
-
-    override func onClick(_ navigationController: UINavigationController?) {
-        navigationController?.dismiss(animated: true) {
-            if let url = URL(string: "https://support.mozilla.org/products/ios") {
-                self.delegate?.settingsOpenURLInNewTab(url)
-            }
-        }
-    }
-}
